+++ src/Ext/House/Body.h
@@ -4,6 +4,8 @@
 #include <Utilities/Container.h>
 #include <Utilities/TemplateDef.h>
 
+#include <array>
+
 class HouseExt
 {
 public:
@@ -69,6 +71,8 @@ public:
 
 		bool PlayerAutoRepair;
 
+		std::array<int, 3> BeaconsPlacedOrder;
+
 		ExtData(HouseClass* OwnerObject) : Extension<HouseClass>(OwnerObject)
 			, PowerPlantEnhancers {}
 			, OwnedLimboDeliveredBuildings {}
@@ -103,6 +107,7 @@ public:
 			, FreeRadar(false)
 			, ForceRadar(false)
 			, PlayerAutoRepair(true)
+			, BeaconsPlacedOrder { 0, 0, 0 }
 		{ }
 
 		bool OwnsLimboDeliveredBuilding(BuildingClass* pBuilding) const;