+++ src/Ext/Rules/Body.h
@@ -379,6 +379,8 @@ public:
 		Valueable<bool> AllowChatBoxInSinglePlayer;
 
 		Valueable<bool> SecondaryFireSequenceLandOnly;
+		Valueable<bool> AutoRemoveEarliestBeacon;
+		Valueable<bool> AllowBeaconHotKeyInSinglePlayer;
 
 		ExtData(RulesClass* OwnerObject) : Extension<RulesClass>(OwnerObject)
 			, Storage_TiberiumIndex { -1 }
@@ -698,6 +700,10 @@ public:
 			, AllowChatBoxInSinglePlayer { false }
 
 			, SecondaryFireSequenceLandOnly { true }
+
+			, AutoRemoveEarliestBeacon { false }
+
+			, AllowBeaconHotKeyInSinglePlayer { false }
 		{ }
 
 		virtual ~ExtData() = default;